Full synchronise on start with initialState
I like this library for it's simplicity, but I stumbled on a frustrating issue. Basically, there is an issue to init state with initial state, but that is pretty much useless as it ignores the values in cookies and localstorage. Instead I needed to run full sync on initialization.

This changes achieves the objective. Now we run `syncUniversal` for every key in `initialState`.

That allows for very neat usage pattern:
For example, if is have an initialState like this in nuxt config:
```
  storage: {
    initialState: { testParam: false }
  },
```


Then in my component (with decorators) I can simply declare
```
  @State(s => s.storage.testParam)
  testParam
```

and on change fire `this.$storage.setUniversal("testParam", newVal)`